+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [Unreleased] - ReleaseDate
 
+## [0.2.1] - 2024-08-02
+
 ### Fixed
 
 - Fix broken internal link in docs